refactor: Extract getProgress function to file
