+++ packages/insight-viewer/src/components/LoadingProgress/index.tsx
@@ -4,6 +4,7 @@ import {
   loadingProgressMessage,
   loadedCountMessageMessage,
 } from '../../utils/messageService'
+import { getProgress } from '../../utils/common/getProgress'
 import { ProgressComponent } from '../../types'
 
 let subscription: Subscription
@@ -15,26 +16,6 @@ const style = {
   transform: 'translateY(-50%)',
   textAlign: 'center',
 } as const
-
-/**
- * Calculate progress status value for single/multiple image(s) fetching.
- * @param {number} loadedCount The count of loaded images.
- * @param {number} totalCount The length of images to load.
- * @param {number} progress The progress status of image to be loading. Repeat 0 to 100 for each new image fetching.
- * @returns {number} Total progress. When all images are loaded, return 100. Otherwise, round down.
- */
-export function getProgress({
-  loadedCount,
-  totalCount,
-  progress,
-}: {
-  loadedCount: number
-  totalCount: number
-  progress: number
-}): number {
-  const calculated = (loadedCount * 100 + (progress ?? 0)) * (1 / totalCount)
-  return loadedCount === totalCount ? 100 : Math.floor(calculated)
-}
 
 export default function LoadingProgress({
   Progress,
